+++ library/src/main/java/candybar/lib/activities/CandyBarMainActivity.java
@@ -152,6 +152,7 @@ public abstract class CandyBarMainActivity extends AppCompatActivity implements
                         R.color.navigationBarDark : R.color.navigationBar));
         if (!Preferences.get(this).isDarkTheme() && (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)) {
             this.getWindow().addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
+            this.getWindow().setStatusBarColor(ContextCompat.getColor(this, R.color.colorPrimaryDark));
             if (ColorHelper.isLightColor(ContextCompat.getColor(this, R.color.navigationBar))
                     && ColorHelper.isLightColor(ContextCompat.getColor(this, R.color.colorPrimaryDark))) {
                 this.getWindow().getDecorView().setSystemUiVisibility(View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR | View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR);
@@ -222,7 +223,7 @@ public abstract class CandyBarMainActivity extends AppCompatActivity implements
             return;
         }
 
-        if (Preferences.get(this).isNewVersion()) {
+        if (!Preferences.get(this).isPlaystoreCheckEnabled() && !mConfig.isLicenseCheckerEnabled() && Preferences.get(this).isNewVersion()) {
             ChangelogFragment.showChangelog(mFragManager);
             File cache = this.getCacheDir();
             FileHelper.clearDirectory(cache);